Move storage to dedicated directory, restyle code to make tslint happy
